Refactored combat_helper and level_up_helper into main_helper
