+++ src/gui/helpers/mod.rs
@@ -1,7 +1,5 @@
-mod combat_helper;
 mod debug_helper;
 mod inventory_helper;
-mod level_up_helper;
 mod main_helper;
 mod route_helper_v1;
 mod shop_helper;
@@ -9,8 +7,7 @@ mod shop_helper;
 use std::collections::HashMap;
 
 use self::{
-    combat_helper::*, debug_helper::*, inventory_helper::*, level_up_helper::*, main_helper::*,
-    route_helper_v1::*, shop_helper::*,
+    debug_helper::*, inventory_helper::*, main_helper::*, route_helper_v1::*, shop_helper::*,
 };
 
 use crate::{game_manager::GameManager, state::GameState};
@@ -26,9 +23,7 @@ impl Default for GuiHelpers {
         helpers.insert(debug_helper::NAME.to_owned(), DebugHelper::create());
         helpers.insert(shop_helper::NAME.to_owned(), ShopHelper::create());
         helpers.insert(inventory_helper::NAME.to_owned(), InventoryHelper::create());
-        helpers.insert(combat_helper::NAME.to_owned(), CombatHelper::create());
         helpers.insert(route_helper_v1::NAME.to_owned(), RouteHelperV1::create());
-        helpers.insert(level_up_helper::NAME.to_owned(), LevelUpHelper::create());
         Self { helpers }
     }
 }
